Initial EMI Support
Adds the relevant calls and a simple structure for EMI.
Registers the workplaces as well, making it a little easier to see all blocks that use a given workstation.
Philosophy for recipe drawing favors showing a input blocks by tag, and then outputs in a 7-wide grid.
Some jank in the current implementation for tags where too many blocks share the same tag, mostly relevant for the Glass recipes by default.